reworking api wrapper
